+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Closing wrong tags in html-only mode should not throw an error
 * Do not render code inside HTML comments
 * Do not parse hawkejs code when setting innerText/innerHTML of an element
+* Fix `Hawkejs.removeChildren(element)` not removing children of `<form>` element
 
 ## 2.2.0 (2021-09-12)